requirements.txt
tensorflow>=2.14.0
numpy
matplotlib
seaborn
streamlit
pillow
scikit-learn
